refactor(health): split HealthPage presentation sections and drift constants
